Move duplicated date logic into component
