+++ app/styles/app.scss
@@ -30,6 +30,7 @@
 @import "loading";
 @import "canary_test_results";
 @import "admin_review";
+@import "components/date-nav";
 
 ::-ms-clear {
   display: none;
