+++ app/styles/components/_date-nav.scss
@@ -1,4 +1,4 @@
-.admin-build-result-nav {
+.date-nav {
   text-align: center;
 
   a.previous-day {
@@ -6,11 +6,5 @@
   }
   a.next-day {
     float: right;
-  }
-}
-
-.admin-build-result {
-  pre {
-    line-height: 1;
   }
 }
