+++ app/templates/components/date-nav.hbs
@@ -1,0 +1,8 @@
+{{yield
+  (hash
+    formattedPreviousDate=formattedPreviousDate
+    formattedDisplayDate=formattedDisplayDate
+    formattedFollowingDate=formattedFollowingDate
+    showFollowingDayLink=showFollowingDayLink
+  )
+}}
